feat: various cleanups
