+++ fendermint/vm/interpreter/src/interpreter.rs
@@ -1,9 +1,4 @@
-use ethers::abi::{Address, Hash};
-use ethers::etherscan::{gas, verify};
-use ethers::providers::Quorum;
-use ipc_api::checkpoint;
 use std::sync::Arc;
-use tendermint::chain;
 use thiserror::Error;
 
 use crate::fvm::state::ipc::GatewayCaller;
@@ -13,41 +8,23 @@ use crate::fvm::upgrades::UpgradeScheduler;
 
 use crate::selector::{select_messages_by_gas_limit, select_messages_until_total_bytes};
 
+use crate::check::check_nonce_and_sufficient_balance;
+use crate::implicit_messages::{execute_cron_message, maybe_push_chain_metadata};
 use crate::types::*;
 
-use crate::{check, check::check_nonce_and_sufficient_balance};
-use crate::{
-    fvm::FvmMessage,
-    signed::{SignedMessageApplyRes, SignedMessageCheckRes, SyntheticMessage},
-    verify::{IllegalMessage, VerifiableMessage},
-    CheckInterpreter, ExecInterpreter, ProposalInterpreter, QueryInterpreter,
-};
+use crate::verify::{IllegalMessage, VerifiableMessage};
 
 use fendermint_vm_message::signed::SignedMessageError;
 
-use anyhow::{anyhow, Chain, Context};
-use fendermint_tracing::emit;
-use fendermint_vm_actor_interface::{chainmetadata, cron, system};
-use fendermint_vm_event::ParentFinalityMissingQuorum;
+use anyhow::Context;
+
 use fendermint_vm_message::chain::ChainMessage;
-use fendermint_vm_message::ipc::ParentFinality;
-use fendermint_vm_message::ipc::{
-    BottomUpCheckpoint, CertifiedMessage, IpcMessage, SignedRelayedMessage,
-};
-use fendermint_vm_resolver::pool::{ResolveKey, ResolvePool};
-use fendermint_vm_topdown::proxy::IPCProviderProxyWithLatency;
-use fendermint_vm_topdown::voting::{ValidatorKey, VoteTally};
-use fendermint_vm_topdown::{
-    CachedFinalityProvider, IPCParentFinality, ParentFinalityProvider, ParentViewProvider, Toggle,
-};
+use fendermint_vm_message::ipc::IpcMessage;
+
 use fvm_ipld_blockstore::Blockstore;
-use fvm_ipld_encoding::RawBytes;
-use fvm_shared::clock::ChainEpoch;
-use fvm_shared::BLOCK_GAS_LIMIT;
 
-pub type TopDownFinalityProvider = Arc<Toggle<CachedFinalityProvider<IPCProviderProxyWithLatency>>>;
-
-use async_stm::{atomically, StmControl};
+use crate::bottomup::BottomUpCheckpointResolver;
+use crate::topdown::TopDownCheckpointResolver;
 
 #[derive(Error, Debug)]
 pub enum InterpreterError {
@@ -59,63 +36,15 @@ pub enum InterpreterError {
     SignedMessageError(#[from] SignedMessageError),
 }
 
-pub type CheckpointPool = ResolvePool<CheckpointPoolItem>;
-
-#[derive(Clone, Hash, PartialEq, Eq)]
-pub enum CheckpointPoolItem {
-    /// BottomUp checkpoints to be resolved from the originating subnet or the current one.
-    BottomUp(CertifiedMessage<BottomUpCheckpoint>),
-    // We can extend this to include top-down checkpoints as well, with slightly
-    // different resolution semantics (resolving it from a trusted parent, and
-    // awaiting finality before declaring it available).
-}
-
-impl From<CertifiedMessage<BottomUpCheckpoint>> for CheckpointPoolItem {
-    fn from(value: CertifiedMessage<BottomUpCheckpoint>) -> Self {
-        CheckpointPoolItem::BottomUp(value)
-    }
-}
-
-impl From<&CheckpointPoolItem> for ResolveKey {
-    fn from(value: &CheckpointPoolItem) -> Self {
-        match value {
-            CheckpointPoolItem::BottomUp(cp) => {
-                (cp.message.subnet_id.clone(), cp.message.bottom_up_messages)
-            }
-        }
-    }
-}
-
-pub enum ProcessDecision {
-    /// The batch of messages meets the criteria and should be included in the block.
-    Accept,
-    /// The batch of messages does not meet the criteria and should be rejected.
-    Reject,
-}
-
-// TODO Karel - handle this type in the check function instead
-// pub enum CheckDecision {
-//     Accept(FvmCheckRet),
-//     Reject,
-// }
-
-// Arbitrarily large gas limit for cron (matching how Forest does it, which matches Lotus).
-// XXX: Our blocks are not necessarily expected to be 30 seconds apart, so the gas limit might be wrong.
-const GAS_LIMIT: u64 = BLOCK_GAS_LIMIT * 10000;
-
 struct Interpreter<DB>
 where
     DB: Blockstore + Clone + 'static + Send + Sync,
 {
-    // TODO Karel - Consider using trait
-    /// TODO Karel - move these into separate module for top down
-    bottom_up_checkpoint_pool: CheckpointPool,
+    bottom_up_resolver: BottomUpCheckpointResolver,
+    top_down_resolver: TopDownCheckpointResolver,
 
     /// The parent finality provider for top down checkpoint
     /// TODO Karel - move these into separate module for top down
-    parent_finality_provider: TopDownFinalityProvider,
-    parent_finality_votes: VoteTally,
-
     // TODO Karel - consider using trait to make this testable
     gateway_caller: GatewayCaller<DB>,
     /// Upgrade scheduler stores all the upgrades to be executed at given heights.
@@ -134,9 +63,8 @@ where
     DB: Blockstore + Clone + 'static + Send + Sync,
 {
     pub fn new(
-        bottom_up_checkpoint_pool: CheckpointPool,
-        parent_finality_provider: TopDownFinalityProvider,
-        parent_finality_votes: VoteTally,
+        bottom_up_resolver: BottomUpCheckpointResolver,
+        top_down_resolver: TopDownCheckpointResolver,
         upgrade_scheduler: UpgradeScheduler<DB>,
         push_chain_meta: bool,
         max_msgs_per_block: usize,
@@ -144,9 +72,8 @@ where
     ) -> Self {
         Self {
             gateway_caller: GatewayCaller::default(),
-            bottom_up_checkpoint_pool,
-            parent_finality_provider,
-            parent_finality_votes,
+            bottom_up_resolver,
+            top_down_resolver,
             upgrade_scheduler,
             push_chain_meta,
             max_msgs_per_block,
@@ -176,7 +103,7 @@ where
         // Check that the signature is valid
         verifiable_msg.verify(&state.chain_id())?;
 
-        let check_ret = check::check_nonce_and_sufficient_balance(&state, &fvm_msg)?;
+        let check_ret = check_nonce_and_sufficient_balance(&state, &fvm_msg)?;
 
         tracing::info!(
             exit_code = check_ret.exit_code.value(),
@@ -221,10 +148,17 @@ where
 
         // Messages generated by the protocol (e.g. top down, bottom up checkpoints)
         // Add bottom up messages ready for execution directly
-        let mut protocol_msgs = self.messages_from_resolved_bottom_up_checkpoints().await;
+        let mut protocol_msgs = self
+            .bottom_up_resolver
+            .messages_from_resolved_checkpoints()
+            .await;
 
-        // Add top down message if exists
-        if let Some(top_down_message) = self.resolve_top_down_exec_message().await {
+        // Add top down message if parent checkpoint is available
+        if let Some(top_down_message) = self
+            .top_down_resolver
+            .resolve_message_from_finality_and_quorum()
+            .await
+        {
             protocol_msgs.push(top_down_message);
         }
 
@@ -299,14 +233,15 @@ where
             match msg {
                 ChainMessage::Ipc(IpcMessage::BottomUpExec(checkpoint)) => {
                     if !self
-                        .check_bottom_up_checkpoint_resolved(checkpoint.into())
+                        .bottom_up_resolver
+                        .check_checkpoint_resolved(checkpoint.into())
                         .await
                     {
                         return Ok(ProcessDecision::Reject);
                     }
                 }
                 ChainMessage::Ipc(IpcMessage::TopDownExec(finality)) => {
-                    if !self.check_parent_checkpoint_finalized(finality).await {
+                    if !self.top_down_resolver.check_valid(finality).await {
                         return Ok(ProcessDecision::Reject);
                     }
                 }
@@ -324,6 +259,27 @@ where
         }
 
         Ok(ProcessDecision::Accept)
+    }
+
+    async fn begin(
+        &self,
+        mut state: FvmExecState<DB>,
+    ) -> anyhow::Result<(FvmExecState<DB>, FvmApplyRet)> {
+        // Block height (FVM epoch) as sequence is intentional
+        let height = state.block_height() as u64;
+
+        // Check for upgrades in the upgrade_scheduler
+        self.maybe_perform_upgrade(&mut state)?;
+
+        // Execute cron message in the cron actor
+        let cron_apply_ret = execute_cron_message(&mut state, height)?;
+
+        // Push the current block hash to the chainmetadata actor if possible
+        if self.push_chain_meta {
+            maybe_push_chain_metadata(&mut state, height)?;
+        }
+
+        Ok((state, cron_apply_ret))
     }
 
     /// Attempts to perform an upgrade if one is scheduled for the current block height,
@@ -345,223 +301,6 @@ where
         }
 
         Ok(())
-    }
-
-    async fn begin(
-        &self,
-        mut state: FvmExecState<DB>,
-    ) -> anyhow::Result<(FvmExecState<DB>, FvmApplyRet)> {
-        // Block height (FVM epoch) as sequence is intentional
-        let height = state.block_height() as u64;
-
-        // Check for upgrades in the upgrade_scheduler
-        self.maybe_perform_upgrade(&mut state)?;
-
-        // Execute cron message in the cron actor
-        let cron_apply_ret = self.execute_cron_message(&mut state, height)?;
-
-        // Push the current block hash to the chainmetadata actor if possible
-        if self.push_chain_meta {
-            self.maybe_push_chain_metadata(&mut state, height)?;
-        }
-
-        Ok((state, cron_apply_ret))
-    }
-
-    async fn check_parent_checkpoint_finalized(&self, msg: ParentFinality) -> bool {
-        let prop = IPCParentFinality {
-            height: msg.height as u64,
-            block_hash: msg.block_hash,
-        };
-        atomically(|| self.parent_finality_provider.check_proposal(&prop)).await
-    }
-
-    async fn check_bottom_up_checkpoint_resolved(
-        &self,
-        msg: CertifiedMessage<BottomUpCheckpoint>,
-    ) -> bool {
-        let item = CheckpointPoolItem::BottomUp(msg);
-
-        // We can just look in memory because when we start the application, we should retrieve any
-        // pending checkpoints (relayed but not executed) from the ledger, so they should be there.
-        // We don't have to validate the checkpoint here, because
-        // 1) we validated it when it was relayed, and
-        // 2) if a validator proposes something invalid, we can make them pay during execution.
-        let is_resolved =
-            atomically(|| match self.bottom_up_checkpoint_pool.get_status(&item)? {
-                None => Ok(false),
-                Some(status) => status.is_resolved(),
-            })
-            .await;
-        is_resolved
-    }
-
-    // Checks the bottom up checkpoint pool and returns the messages that are ready for execution
-    async fn messages_from_resolved_bottom_up_checkpoints(&self) -> Vec<ChainMessage> {
-        let resolved = atomically(|| self.bottom_up_checkpoint_pool.collect_resolved()).await;
-        resolved
-            .into_iter()
-            .map(|checkpoint| match checkpoint {
-                CheckpointPoolItem::BottomUp(checkpoint) => {
-                    ChainMessage::Ipc(IpcMessage::BottomUpExec(checkpoint))
-                }
-            })
-            .collect()
-    }
-
-    /// Prepares a top-down execution message based on the current parent's finality proposal and quorum.
-    ///
-    /// This function first pauses incoming votes to prevent interference during processing. It then atomically retrieves
-    /// both the next parent's proposal and the quorum of votes. If either the parent's proposal or the quorum is missing,
-    /// the function returns `None`. When both are available, it selects the finality with the lower block height and wraps
-    /// it into a `ChainMessage` for top-down execution.
-    async fn resolve_top_down_exec_message(&self) -> Option<ChainMessage> {
-        // Prepare top down proposals.
-        // Before we try to find a quorum, pause incoming votes. This is optional but if there are lots of votes coming in it might hold up proposals.
-        atomically(|| self.parent_finality_votes.pause_votes_until_find_quorum()).await;
-
-        // The pre-requisite for proposal is that there is a quorum of gossiped votes at that height.
-        // The final proposal can be at most as high as the quorum, but can be less if we have already,
-        // hit some limits such as how many blocks we can propose in a single step.
-        let (parent, quorum) = atomically(|| {
-            let parent = self.parent_finality_provider.next_proposal()?;
-
-            let quorum = self
-                .parent_finality_votes
-                .find_quorum()?
-                .map(|(height, block_hash)| IPCParentFinality { height, block_hash });
-
-            Ok((parent, quorum))
-        })
-        .await;
-
-        // If there is no parent proposal, exit early.
-        let parent = if let Some(parent) = parent {
-            parent
-        } else {
-            return None;
-        };
-
-        // Require a quorum; if it's missing, log and exit.
-        let quorum = if let Some(quorum) = quorum {
-            quorum
-        } else {
-            emit!(
-                DEBUG,
-                ParentFinalityMissingQuorum {
-                    block_height: parent.height,
-                    block_hash: &hex::encode(&parent.block_hash),
-                }
-            );
-            return None;
-        };
-
-        // Choose the lower height between the parent's proposal and the quorum.
-        let finality = if parent.height <= quorum.height {
-            parent
-        } else {
-            quorum
-        };
-
-        Some(ChainMessage::Ipc(IpcMessage::TopDownExec(ParentFinality {
-            height: finality.height as ChainEpoch,
-            block_hash: finality.block_hash,
-        })))
-    }
-
-    /// Executes the cron message for the given block height.
-    fn execute_cron_message(
-        &self,
-        state: &mut FvmExecState<DB>,
-        height: u64,
-    ) -> anyhow::Result<FvmApplyRet> {
-        let from = system::SYSTEM_ACTOR_ADDR;
-        let to = cron::CRON_ACTOR_ADDR;
-        let method_num = cron::Method::EpochTick as u64;
-        let gas_limit = GAS_LIMIT;
-
-        let msg = FvmMessage {
-            from,
-            to,
-            sequence: height,
-            gas_limit,
-            method_num,
-            params: Default::default(),
-            value: Default::default(),
-            version: Default::default(),
-            gas_fee_cap: Default::default(),
-            gas_premium: Default::default(),
-        };
-
-        let (apply_ret, emitters) = state.execute_implicit(msg)?;
-
-        if let Some(err) = apply_ret.failure_info {
-            anyhow::bail!("failed to apply block cron message: {}", err);
-        }
-
-        Ok(FvmApplyRet {
-            apply_ret,
-            emitters,
-            from,
-            to,
-            method_num,
-            gas_limit,
-        })
-    }
-
-    /// Attempts to push chain metadata if a block hash is available.
-    fn maybe_push_chain_metadata(
-        &self,
-        state: &mut FvmExecState<DB>,
-        height: u64,
-    ) -> anyhow::Result<Option<FvmApplyRet>> {
-        let gas_limit = GAS_LIMIT;
-        let from = system::SYSTEM_ACTOR_ADDR;
-        let to = chainmetadata::CHAINMETADATA_ACTOR_ADDR;
-        let method_num = fendermint_actor_chainmetadata::Method::PushBlockHash as u64;
-
-        // Proceed only if the current block has a hash.
-        if let Some(block_hash) = state.block_hash() {
-            // Serialize the push parameters.
-            let params = fvm_ipld_encoding::RawBytes::serialize(
-                fendermint_actor_chainmetadata::PushBlockParams {
-                    // TODO: this conversion from u64 to i64 should be revisited.
-                    epoch: height as i64,
-                    block: block_hash,
-                },
-            )?;
-
-            let msg = FvmMessage {
-                from,
-                to,
-                sequence: height,
-                gas_limit,
-                method_num,
-                params,
-                value: Default::default(),
-                version: Default::default(),
-                gas_fee_cap: Default::default(),
-                gas_premium: Default::default(),
-            };
-
-            let (apply_ret, emitters) = state.execute_implicit(msg)?;
-            if let Some(err) = apply_ret.failure_info {
-                anyhow::bail!("failed to apply chainmetadata message: {}", err);
-            }
-
-            let fvm_apply_ret = FvmApplyRet {
-                apply_ret,
-                emitters,
-                from,
-                to,
-                method_num,
-                gas_limit,
-            };
-            Ok(Some(fvm_apply_ret))
-        } else {
-            // No block hash available; nothing to push.
-            Ok(None)
-        }
     }
 }
