+++ fendermint/vm/interpreter/src/lib.rs
@@ -2,12 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 use async_trait::async_trait;
 
+mod bottomup;
 pub mod bytes;
 pub mod chain;
 mod check;
 pub mod fvm;
 pub mod genesis;
+mod implicit_messages;
 pub mod signed;
+mod topdown;
 mod verify;
 
 #[cfg(feature = "arb")]
