+++ fendermint/vm/interpreter/src/topdown.rs
@@ -1,0 +1,91 @@
+use async_stm::atomically;
+use fendermint_tracing::emit;
+use fendermint_vm_message::ipc::IpcMessage;
+use fendermint_vm_topdown::proxy::IPCProviderProxyWithLatency;
+use fendermint_vm_topdown::voting::VoteTally;
+use fendermint_vm_topdown::{
+    CachedFinalityProvider, IPCParentFinality, ParentFinalityProvider, Toggle,
+};
+use fvm_shared::clock::ChainEpoch;
+use std::sync::Arc;
+
+use fendermint_vm_event::ParentFinalityMissingQuorum;
+use fendermint_vm_message::chain::ChainMessage;
+use fendermint_vm_message::ipc::ParentFinality;
+
+type TopDownFinalityProvider = Arc<Toggle<CachedFinalityProvider<IPCProviderProxyWithLatency>>>;
+
+pub struct TopDownCheckpointResolver {
+    provider: TopDownFinalityProvider,
+    votes: VoteTally,
+}
+
+impl TopDownCheckpointResolver {
+    pub async fn check_valid(&self, finality: ParentFinality) -> bool {
+        let prop = IPCParentFinality {
+            height: finality.height as u64,
+            block_hash: finality.block_hash,
+        };
+        atomically(|| self.provider.check_proposal(&prop)).await
+    }
+
+    /// Prepares a top-down execution message based on the current parent's finality proposal and quorum.
+    ///
+    /// This function first pauses incoming votes to prevent interference during processing. It then atomically retrieves
+    /// both the next parent's proposal and the quorum of votes. If either the parent's proposal or the quorum is missing,
+    /// the function returns `None`. When both are available, it selects the finality with the lower block height and wraps
+    /// it into a `ChainMessage` for top-down execution.
+    pub async fn resolve_message_from_finality_and_quorum(&self) -> Option<ChainMessage> {
+        // Prepare top down proposals.
+        // Before we try to find a quorum, pause incoming votes. This is optional but if there are lots of votes coming in it might hold up proposals.
+        atomically(|| self.votes.pause_votes_until_find_quorum()).await;
+
+        // The pre-requisite for proposal is that there is a quorum of gossiped votes at that height.
+        // The final proposal can be at most as high as the quorum, but can be less if we have already,
+        // hit some limits such as how many blocks we can propose in a single step.
+        let (parent, quorum) = atomically(|| {
+            let parent = self.provider.next_proposal()?;
+
+            let quorum = self
+                .votes
+                .find_quorum()?
+                .map(|(height, block_hash)| IPCParentFinality { height, block_hash });
+
+            Ok((parent, quorum))
+        })
+        .await;
+
+        // If there is no parent proposal, exit early.
+        let parent = if let Some(parent) = parent {
+            parent
+        } else {
+            return None;
+        };
+
+        // Require a quorum; if it's missing, log and exit.
+        let quorum = if let Some(quorum) = quorum {
+            quorum
+        } else {
+            emit!(
+                DEBUG,
+                ParentFinalityMissingQuorum {
+                    block_height: parent.height,
+                    block_hash: &hex::encode(&parent.block_hash),
+                }
+            );
+            return None;
+        };
+
+        // Choose the lower height between the parent's proposal and the quorum.
+        let finality = if parent.height <= quorum.height {
+            parent
+        } else {
+            quorum
+        };
+
+        Some(ChainMessage::Ipc(IpcMessage::TopDownExec(ParentFinality {
+            height: finality.height as ChainEpoch,
+            block_hash: finality.block_hash,
+        })))
+    }
+}
